server: Move UseSpecificHttpStatusCodes to rule package #344
